requirements.txt
fal-client>=0.5.0
Pillow>=10.0.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
lpips>=0.1.4
python-dotenv>=1.0.0
tqdm>=4.65.0
matplotlib>=3.7.0
scipy>=1.10.0
jupyter>=1.0.0
google-cloud-aiplatform>=1.38.0
requests>=2.31.0
